pyproject.toml
[tool.poetry]
name = "market-research-agent"
version = "0.1.0"
description = ""
authors = ["Archit <architojha420@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
duckduckgo-search-api = "^0.1.5"
duckduckgo-search = "^7.1.1"
httpx = "0.27.2"
pandas = "^2.2.3"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pydantic-ai = "^0.0.15"
nest-asyncio = "^1.6.0"
devtools = "^0.12.2"
langchain-ollama = "^0.2.2"
llama-index = "^0.12.8"
llama-index-llms-ollama = "^0.5.0"
llama-index-llms-groq = "^0.3.1"
pydantic-settings = "^2.7.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
